BatePontoSanMarinho123/face_service/requirements.txt
fastapi
uvicorn
python-multipart
psycopg2
numpy
pillow
face_recognition
